requirements.txt
click==6.7
dateutils==0.6.6
fdb==1.7
Flask==0.12.2
Flask-Markdown==0.3
flask-simplelogin==0.0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
pymongo==3.6.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
WTForms==2.1